refactor: move facade to service and support Related to: #128
