+++ src/main/java/com/komentum/post/mapper/DesignBoardMapperSupport.java
@@ -1,6 +1,9 @@
 package com.komentum.post.mapper;
 
 import com.komentum.global.utils.DateUtils;
+import com.komentum.designcomponent.domain.ComponentType;
+import com.komentum.designcomponent.enums.TypeCode;
+import com.komentum.post.domain.DesignBoard;
 import com.komentum.post.domain.Tag;
 import com.komentum.post.dto.BoardComponentTypeDto;
 import com.komentum.post.dto.DesignBoardDto.DesignBoardDetailDto;
@@ -8,11 +11,24 @@ import com.komentum.post.dto.DesignBoardDto.DesignBoardPreviewDto;
 import com.komentum.post.dto.TagDto.TagResponse;
 import com.komentum.post.dto.query.DesignBoardQuery;
 import com.komentum.post.facade.BoardManagementHelper;
+import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
 import org.springframework.stereotype.Component;
 
 @Component
 public class DesignBoardMapperSupport {
+
+  public List<BoardComponentTypeDto> toComponentTypeDtos(List<DesignBoard> designBoards) {
+    Map<TypeCode, BoardComponentTypeDto> componentTypeMap = new LinkedHashMap<>();
+    for (DesignBoard designBoard : designBoards) {
+      for (ComponentType componentType : designBoard.getDesignComponent().getComponentTypes()) {
+        componentTypeMap.putIfAbsent(componentType.getTypeCode(),
+            BoardComponentTypeDto.from(componentType));
+      }
+    }
+    return List.copyOf(componentTypeMap.values());
+  }
 
   public DesignBoardPreviewDto toDesignBoardPreviewDto(
       DesignBoardQuery.Preview preview,
