+++ src/main/java/com/komentum/post/service/DesignBoardService.java
@@ -14,6 +14,8 @@ import com.komentum.user.domain.User;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -58,6 +60,16 @@ public class DesignBoardService {
   @Transactional(readOnly = true)
   public List<DesignBoard> findWithDesignComponentsByPostIdIn(List<Long> postIds) {
     return designBoardRepository.findWithDesignComponentByPost_PostIdIn(postIds);
+  }
+
+  @Transactional(readOnly = true)
+  public Map<Long, List<DesignBoard>> findWithDesignComponentsByPostIdMap(List<Long> postIds) {
+    if (postIds == null || postIds.isEmpty()) {
+      return Map.of();
+    }
+    return findWithDesignComponentsByPostIdIn(postIds)
+        .stream()
+        .collect(Collectors.groupingBy(designBoard -> designBoard.getPost().getPostId()));
   }
 
   @Transactional(readOnly = true)
