+++ src/main/java/com/komentum/post/service/transaction/DesignBoardTransactionService.java
@@ -1,14 +1,11 @@
 package com.komentum.post.service.transaction;
 
-import com.komentum.designcomponent.domain.ComponentType;
 import com.komentum.designcomponent.domain.DesignComponent;
-import com.komentum.designcomponent.enums.TypeCode;
 import com.komentum.designcomponent.service.DesignComponentService;
 import com.komentum.post.domain.DesignBoard;
 import com.komentum.post.domain.Post;
 import com.komentum.post.domain.Tag;
 import com.komentum.post.domain.enums.PostType;
-import com.komentum.post.dto.BoardComponentTypeDto;
 import com.komentum.post.dto.DesignBoardDto.DesignBoardCreateDto;
 import com.komentum.post.dto.DesignBoardDto.DesignBoardDetailDto;
 import com.komentum.post.dto.DesignBoardDto.DesignBoardUpdateDto;
@@ -26,9 +23,7 @@ import com.komentum.post.service.TagService;
 import com.komentum.user.domain.User;
 import com.komentum.user.service.UserEntityFinder;
 import jakarta.persistence.EntityNotFoundException;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Stream;
 import lombok.RequiredArgsConstructor;
@@ -78,18 +73,7 @@ public class DesignBoardTransactionService {
         detail,
         tags,
         previewImageUrls,
-        toComponentTypeDtos(designBoards));
-  }
-
-  private List<BoardComponentTypeDto> toComponentTypeDtos(List<DesignBoard> designBoards) {
-    Map<TypeCode, BoardComponentTypeDto> componentTypeMap = new LinkedHashMap<>();
-    for (DesignBoard designBoard : designBoards) {
-      for (ComponentType componentType : designBoard.getDesignComponent().getComponentTypes()) {
-        componentTypeMap.putIfAbsent(componentType.getTypeCode(),
-            BoardComponentTypeDto.from(componentType));
-      }
-    }
-    return List.copyOf(componentTypeMap.values());
+        designBoardMapperSupport.toComponentTypeDtos(designBoards));
   }
 
   @Transactional
